Fixed balance calculation

+++ src/main/java/com/bernardomg/association/balance/model/Balance.java
@@ -3,6 +3,6 @@ package com.bernardomg.association.balance.model;
 
 public interface Balance {
 
-    public Long getAmount();
+    public Float getAmount();
 
 }
