+++ src/main/java/com/bernardomg/association/balance/model/DtoBalance.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 public class DtoBalance implements Balance {
 
-    private Long amount;
+    private Float amount;
 
 }
